Add files via upload
CONTENTS: 

1. Jupyter Source File "Report"
2. CSV File "SeoulBikeData_mov"
3. PDF File "SeoulBikePaper" (This is the research article that the report is based on.)

DESCRIPTION:

This folder contains a short report, based on the 2020 research article by Sathishkumar V E, Jangwoo Park and Yongyun Cho in the journal "Computer Communications".

The article aims to use machine learning and data mining methods to predict the demand for shared bikes in metropolitan cities subject to specific parameters, the most notable of which are: time of day, month, and weather conditions.

In this report we focus on two such methods: Linear Regression and Regression Trees. We read the data and use it to first plot several statistical diagrams, including Histograms, Scatter Diagrams and Violin Plots. We then compute four performance measures for each method, and use them to draw conclusions about the accuracy of our predictive models.